Fix local ci error && refactor log subscription namespace
