+++ scripts/install_deps.sh
@@ -524,6 +524,9 @@ install_cppkafka() {
     return 0
   fi
 
+  export LDFLAGS="-L/usr/local/opt/openssl@3/lib"
+  export CPPFLAGS="-I/usr/local/opt/openssl@3/include"
+
   check_and_remove_dir "/tmp/cppkafka"
   git clone -b 0.4.0 --single-branch --depth=1 \
       https://github.com/mfontanini/cppkafka.git /tmp/cppkafka
